# docs/swagger.yaml
definitions:
  entity.RoleMenu:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      menu_id:
        type: string
      role_id:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
  entity.RolePermission:
    properties:
      id:
        type: string
      permission_id:
        type: string
      role_id:
        type: string
    type: object
...
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  request.RoleHasMenu:
    properties:
      menus:
        items:
          $ref: '#/definitions/entity.RoleMenu'
        type: array
    type: object
  request.RoleHasPermission:
    properties:
      permissions:
        items:
          $ref: '#/definitions/entity.RolePermission'
        type: array
    type: object
  request.RoleRequest:
    properties:
      created_by:
        type: string
      description:
        type: string
      name:
        type: string
      updated_by:
        type: string
    required:
    - name
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Api Service boilerplate dari golang gin dan gorm
  termsOfService: http://swagger.io/terms/
  title: My Admin Service Api
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: REST API for Roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get list Roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: REST API for Roles
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Insert Roles
      tags:
      - Roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: REST API for Roles
      parameters:
      - description: Roles ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Delete Roles
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: REST API for Roles
      parameters:
      - description: Roles ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Get one Roles
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: REST API for Roles
      parameters:
      - description: Roles ID
        in: query
        name: id
        type: string
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Update Roles
      tags:
      - Roles
  /roles/menu:
    delete:
      consumes:
      - application/json
      description: REST API Role Menu
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RoleHasMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Detach role has menus
      tags:
      - Roles
  /roles/menus:
    post:
      consumes:
      - application/json
      description: REST API Role Menu
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RoleHasMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Attach role has menu
      tags:
      - Roles
  /roles/permissions:
    delete:
      consumes:
      - application/json
      description: REST API Role Permission
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RoleHasPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Detach role has permissions
      tags:
      - Roles
      - Roles
    post:
      consumes:
      - application/json
      description: REST API Role Permission
      parameters:
      - description: Input data in JSON format
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RoleHasPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerAuth: []
      summary: Attach role has permissions
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"